requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==1.9
django-environ==0.4.5
idna==2.10
inflection==0.5.1
oauthlib==3.1.0
profanityfilter==2.0.6
psycopg2-binary
PySocks==1.7.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.9.0
urllib3==1.25.10
